move keycloak out of utils to have its own module
